pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "optexity"
version = "0.1.0"
readme = "README.md"
description = "Optexity is a platform for building and running browser and computer agents."
authors = [{ name = "Optexity", email = "support@optexity.com" }]
requires-python = ">=3.11"
dependencies = [
    "pydantic_settings",
    "pydantic",
    "playwright",
    "patchright",
    "tokencost",
    "pre-commit",
    "black",
    "isort",
    "aiofiles",
    "onepassword-sdk"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["optexity"]

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true